AI Deployment/cat-dog-classifier/app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.1
torch==2.2.0
torchvision==0.17.0
prometheus-fastapi-instrumentator==7.0.2
python-multipart==0.0.7
Pillow==10.2.0